change entities to "entities" folder
