+++ backend/src/entities/manufacturers/manufacturers.controller.ts
