+++ backend/src/entities/manufacturers/manufacturers.module.ts
