+++ backend/src/entities/manufacturers/manufacturers.service.ts
