+++ backend/src/entities/orders/orders.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { OrdersService } from './orders.service';
 import { OrdersController } from './orders.controller';
-import { UsersModule } from 'src/users/users.module';
+import { UsersModule } from 'src/entities/users/users.module';
 import { PrismaModule } from 'src/prisma/prisma.module';
 
 @Module({
