+++ backend/src/entities/part-templates/part-templates.module.ts
