+++ backend/src/entities/part-templates/part-templates.service.ts
