+++ backend/src/entities/users/users.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
 import { PrismaModule } from 'src/prisma/prisma.module';
-import { ContactsModule } from 'src/contacts/contacts.module';
-import { AddressesModule } from 'src/addresses/addresses.module';
+import { ContactsModule } from 'src/entities/contacts/contacts.module';
+import { AddressesModule } from 'src/entities/addresses/addresses.module';
 
 @Module({
   imports: [PrismaModule, ContactsModule, AddressesModule],
